Update from libmemcached's libtest

+++ libtest/server.cc
@@ -96,7 +96,8 @@ Server::~Server()
 bool Server::check()
 {
   _app.slurp();
-  return _app.check();
+  _app.check();
+  return true;
 }
 
 bool Server::validate()
@@ -188,18 +189,30 @@ bool Server::start()
     dream(5, 50000);
   }
 
-  if (pid_file().empty() == false)
+  size_t repeat= 5;
+  while (--repeat)
   {
-    Wait wait(pid_file(), 20);
-
-    if (wait.successful() == false)
+    if (pid_file().empty() == false)
     {
-      char buf[PATH_MAX];
-      getcwd(buf, sizeof(buf));
-      throw libtest::fatal(LIBYATL_DEFAULT_PARAM,
-                           "Unable to open pidfile in %s for: %s",
-                           buf,
-                           _running.c_str());
+      Wait wait(pid_file(), 8);
+
+      if (wait.successful() == false)
+      {
+        if (_app.check())
+        {
+          continue;
+        }
+
+        throw libtest::fatal(LIBYATL_DEFAULT_PARAM,
+                             "Unable to open pidfile for: %s",
+                             _running.c_str());
+        char buf[PATH_MAX];
+        getcwd(buf, sizeof(buf));
+        throw libtest::fatal(LIBYATL_DEFAULT_PARAM,
+                             "Unable to open pidfile in %s for: %s",
+                             buf,
+                             _running.c_str());
+      }
     }
   }
 
@@ -229,19 +242,22 @@ bool Server::start()
   if (pinged == false)
   {
     // If we happen to have a pid file, lets try to kill it
-    if (pid_file().empty() == false)
+    if ((pid_file().empty() == false) and (access(pid_file().c_str(), R_OK) == 0))
     {
       if (kill_file(pid_file()) == false)
       {
         throw libtest::fatal(LIBYATL_DEFAULT_PARAM, "Failed to kill off server after startup occurred, when pinging failed: %s", pid_file().c_str());
       }
-      Error << "Failed to ping(), waited:" << this_wait 
-        << " server started, having pid_file. exec:" << _running 
-        << " error:" << _app.stderr_result();
+
+      throw libtest::fatal(LIBYATL_DEFAULT_PARAM, 
+                           "Failed to ping(), waited: %u server started, having pid_file. exec: %s error:%s",
+                           this_wait, _running.c_str(), _app.stderr_c_str()); 
     }
     else
     {
-      Error << "Failed to ping() server started. exec:" << _running;
+      throw libtest::fatal(LIBYATL_DEFAULT_PARAM,
+                           "Failed to ping() server started. exec: %s",
+                           _running.c_str());
     }
     _running.clear();
     return false;
@@ -348,7 +364,7 @@ bool Server::args(Application& app)
 {
 
   // Set a log file if it was requested (and we can)
-  if (has_log_file_option())
+  if (false and has_log_file_option())
   {
     set_log_file();
     log_file_option(app, _log_file);
